src/ch5/requirements.txt
scikit-learn==1.4.2
pandas==2.1.1
numpy==1.26.3
matplotlib==3.8.0
matplotlib_fontja==1.0.0
tensorflow==2.16.1
keras==3.2.1
Flask==3.0.3
Pillow==10.3.0
